Reduce repetition by extracting FormattingDelegate/DateFormatterDelegate.
